[add,mod,del] removes annotation index from store  - moves pointer functions to separate hook for clarity  - removes annotation state logic from selection tool
