+++ src/hooks/usePointerTool/index.ts
@@ -1,0 +1,1 @@
+export { usePointerTool } from "./usePointerTool";
