--- src/store/imageViewer/selectors/currentIndexSelector.ts
@@ -1,8 +1,0 @@
-import { ImageViewerStore } from "types";
-export const currentIndexSelector = ({
-  imageViewer,
-}: {
-  imageViewer: ImageViewerStore;
-}) => {
-  return imageViewer.currentIndex;
-};
